# PaymentSolidApi/src/main/resources/api/openapi.yml
openapi: 3.0.3
info:
  title: API de Gerenciamento de Pagamentos
  description: API para processar e gerenciar pagamentos através de diferentes métodos (Cartão de Crédito, PIX, Boleto).
  version: 1.0.0
servers:
  - url: http://localhost:8082/api/v1
    description: Servidor de Desenvolvimento
tags:
  - name: Payments
    description: Operações relacionadas ao processamento e ciclo de vida de pagamentos
paths:
  /payments:
    post:
      tags:
        - Payments
      summary: Criar Novo Pagamento
      description: Processa um novo pagamento no sistema.
      operationId: createPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequestDTO'
            examples:
              credit_card:
                summary: Pagamento com Cartão de Crédito
                value:
                  paymentMethod: CREDIT_CARD
                  amount: 150.75
                  cardNumber: '1234567890123456'
                  cvv: '123'
              pix:
                summary: Pagamento com PIX
                value:
                  paymentMethod: PIX
                  amount: 89.90
                  pixKey: 'email@example.com'
              boleto:
                summary: Pagamento com Boleto
                value:
                  paymentMethod: BOLETO
                  amount: 250.00
      responses:
        '201':
          description: Pagamento criado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponseDTO'
        '400':
          description: Dados de pagamento inválidos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        '500':
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'

  /payments/{id}:
    get:
      tags:
        - Payments
      summary: Buscar um Pagamento
      description: Busca um pagamento no sistema pelo seu ID único.
      operationId: findPaymentById
      parameters:
        - name: id
          in: path
          required: true
          description: ID único do pagamento
          schema:
            type: string
            format: uuid
            example: 'a1b2c3d4-e5f6-7890-1234-567890abcdef'
      responses:
        '200':
          description: Pagamento encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponseDTO'
        '404':
          description: Pagamento não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        '500':
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'

  /payments/refund/{id}:
    post:
      tags:
        - Payments
      summary: Estornar Pagamento
      description: |
        Realiza o estorno de um pagamento.
        - **PIX:** Não permite estorno automático.
        - **Boleto:** Precisa de cancelamento antes do vencimento.
        - **Cartão de Crédito:** Permite estorno total.
      operationId: refundPayment
      parameters:
        - name: id
          in: path
          required: true
          description: ID único do pagamento a ser estornado
          schema:
            type: string
            format: uuid
            example: 'a1b2c3d4-e5f6-7890-1234-567890abcdef'
      responses:
        '200':
          description: Pagamento estornado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponseDTO'
        '400':
          description: Pagamento não pode ser estornado (ex regra de negócio).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        '409':
          description: Conflito de estado ou regra de negócio para estorno.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        '404':
          description: Pagamento não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        '500':
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'

components:
  schemas:
    PaymentMethodEnum:
      type: string
      description: Método de pagamento
      enum:
        - CREDIT_CARD
        - PIX
        - BOLETO
      example: "CREDIT_CARD"

    StatusEnum:
      type: string
      description: Status atual do pagamento
      enum:
        - PENDING
        - PROCESSING
        - APPROVED
        - REJECTED
        - REFUNDED
      example: "APPROVED"

    PaymentRequestDTO:
      type: object
      description: Dados para processar um novo pagamento
      required:
        - paymentMethod
        - amount
      properties:
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethodEnum'
        amount:
          type: number
          format: decimal
          description: Valor do pagamento
          minimum: 0.01
          example: 150.75
        cardNumber:
          type: string
          description: 'Número do cartão de crédito (obrigatório se paymentMethod for CREDIT_CARD).'
          pattern: '^[0-9]{16}$'
          example: '1234567890123456'
        cvv:
          type: string
          description: 'CVV do cartão de crédito (obrigatório se paymentMethod for CREDIT_CARD).'
          pattern: '^[0-9]{3,4}$'
          example: '123'
        pixKey:
          type: string
          description: "Chave PIX (obrigatório se paymentMethod for PIX)."
          example: 'email@example.com'

    PaymentResponseDTO:
      type: object
      description: Dados retornados após o processamento de um pagamento
      properties:
        paymentId:
          type: string
          format: uuid
          description: ID único do pagamento gerado
          example: 'a1b2c3d4-e5f6-7890-1234-567890abcdef'
        status:
          $ref: '#/components/schemas/StatusEnum'
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethodEnum'
        amount:
          type: number
          format: decimal
          description: Valor do pagamento processado
          example: 150.75
        message:
          type: string
          description: Mensagem de confirmação ou status
          example: 'Pagamento processado com sucesso.'
        processedAt:
          type: string
          format: date-time
          description: Data e hora em que o pagamento foi processado
          example: '2025-09-13T18:45:00Z'

    ErrorResponseDTO:
      type: object
      properties:
        code:
          type: string
          description: Código estável para classificação do erro
          example: "PAYMENT_NOT_FOUND"
        message:
          type: string
          description: Descrição do erro
          example: "Pagamento não encontrado"
        details:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          description: Detalhes adicionais do erro (opcional)
        traceId:
          type: string
          description: Identificador para correlação e troubleshooting
          example: "a4f8d104-b583-4f15-a585-9474f43bd4b7"